Add StoreInfo class
StoreInfo is intended to be a dictionary of zarr storage parameters
with additional parsing and preprocessing methods. For now, it is
intented to only support zarr arrays saved to disk, though support
for other types of arrays may be added in the future. The initial
version committed here is very simple and will likely change as
additional functionality is added.

This commit also introduces helper init and parsing functions.