Make 'store' an un-settable item for StoreInfo
This commit changes how the 'store' item is handled by StoreInfo.
Currently, 'path' and 'identifier' can be set and 'store' is dynamically
created when calling __getitem__('store') or __getattr__('store').
However, this means that 'store' will not be returned when unpacking a
StoreInfo instance using **.

This commit makes 'store' a regular key in StoreInfo. However, 'store'
will not be able to be set directly. Updating 'path' or 'identifier'
will update 'store' if possible, depending on the store_type. This
change requires additional changes to ModelStoreInfo, helper functions,
and unit tests that are also added in this commit.