Add initial load logic
Add store_load.load, update __init__.py to import load, and add
load_metadata method to ModelStoreInfo.

store_load.load takes an initialized model and replaces each parameter
and buffer with previously stored DriveTensors. This implementation
requires the initialization of an entire model and may not work for
very large models on systems with limited memory.

Loading will be further modified in a future commit to ignore tensor
data when initializing the model in order to reduce the memory
footprint.